2009-10-22  Hidehisa Akiyama  <akky@users.sourceforge.jp>
	* src/field.cpp:
	* src/field.h:
	* src/pcombuilder.h:
	* src/player.cpp:
	* src/player.h:
	* src/player_command_parser.ypp:
	* src/player_command_tok.lpp:
	* src/referee.cpp:
	* src/referee.h:
	* src/serverparam.cpp:
	* src/serverparam.h:
	* src/types.h:
	- implemented new intentional foul model.
	- implemented yellow/red card and discard of red carded player.



git-svn-id: file:///home/akiyama/sserver/svn/rcss/trunk/rcssserver@502 1639d2a8-b03b-4162-8b0b-03f2cde17881

+++ src/serverparam.cpp
@@ -348,7 +348,7 @@ const double ServerParam::MIN_DASH_POWER = -100.0;
 // 14.0.0
 const double ServerParam::TACKLE_RAND_FACTOR = 2.0;
 const double ServerParam::FOUL_DETECT_PROBABILITY = 0.5;
-const int ServerParam::FOUL_EXPONENT = 10;
+const double ServerParam::FOUL_EXPONENT = 10.0;
 const int ServerParam::FOUL_CYCLES = 5;
